Store the creator of a connection and the connection itself in ConnectionInformation
That way, JdbcEventListeners have access for example to the MetaData of the Connection
to associate the execution with a Database URL. JdbcEventListeners can also reason about
how the connection was obtained (Driver/DataSource/PooledConnection) and how long
it took to get the connection. High connection times can for example indicate a exhausted
connection pool.